:recycle: Moved conversion functions to dedicated module.
